renamed ModalButton to ModalButtonWithRouter, renamed GlossaryModal to ModalButton
