+++ client/src/components/modals_and_popovers/ModalButton.tsx
@@ -1,74 +1,42 @@
-import React, { Fragment } from "react";
-
-import { withRouter, RouteComponentProps } from "react-router-dom";
-
-import { SafeJSURL } from "src/general_utils";
-
-import { secondaryColor } from "src/style_constants/index";
+import React from "react";
 
 import { StatelessModal } from "./StatelessModal";
 
-type ModalButtonProps = RouteComponentProps<{ options: string }> & {
-  show_condition: { name: string; value: string };
-  button_text: string;
-  title: string;
-  children: React.ReactNode;
-  aria_label: string;
+type ModalButtonState = {
+  show_modal: boolean;
 };
 
-interface ModalButtonState {
-  show_modal: boolean;
-}
+type ModalButtonProps = {
+  text: string;
+  body: React.ReactNode;
+};
 
-class ModalButton_ extends React.Component<ModalButtonProps, ModalButtonState> {
-  constructor(props: ModalButtonProps) {
-    super(props);
-
-    const {
-      show_condition,
-      match: {
-        params: { options },
-      },
-    } = this.props;
-
-    const options_object = SafeJSURL.parse(options);
-
-    this.state = {
-      show_modal:
-        show_condition &&
-        options_object[show_condition.name] &&
-        show_condition.value === options_object[show_condition.name],
-    };
-  }
+export class ModalButton extends React.Component<
+  ModalButtonProps,
+  ModalButtonState
+> {
+  state = { show_modal: true };
 
   toggle_modal = () =>
     this.setState((prev_state) => ({ show_modal: !prev_state.show_modal }));
 
   render() {
-    const { button_text, title, children, aria_label } = this.props;
+    const { text, body } = this.props;
     const { show_modal } = this.state;
+    console.log("\nModalButton - render");
+    console.log(this.props);
+    console.log(this.state);
 
     return (
-      <Fragment>
-        <button
-          className="btn btn-link"
-          onClick={this.toggle_modal}
-          aria-label={aria_label}
-          style={{ color: secondaryColor }}
-        >
-          {button_text}
-        </button>
+      <>
+        <button onClick={this.toggle_modal}>{text}</button>
         <StatelessModal
           show={show_modal}
-          title={title}
-          body={children}
+          size="sm"
+          body={body}
           on_close_callback={this.toggle_modal}
-          additional_dialog_class={"modal-responsive"}
         />
-      </Fragment>
+      </>
     );
   }
 }
-
-const ModalButton = withRouter(ModalButton_);
-export { ModalButton };
