+++ client/src/components/modals_and_popovers/index.js
@@ -1,4 +1,4 @@
 export { StatelessModal } from "./StatelessModal";
 export { FixedPopover } from "./FixedPopover";
+export { ModalButtonWithRouter } from "./ModalButtonWithRouter";
 export { ModalButton } from "./ModalButton";
-export { GlossaryModal } from "./GlossaryModal";
